requirements.txt
faiss-cpu
numpy
langchain
langchain-google-genai
sentence-transformers
python-dotenv
